+++ NEWS.md
@@ -5,6 +5,9 @@
 * New `{.num}` and `{.bytes}` inline styles to format numbers
   and bytes (@m-muecke, #644, #588, #643).
 
+* Make `pluralize()` support values exceeding `.Machine$integer.max` (@mcol,
+  #773).
+
 # cli 3.6.5
 
 * `code_highlight()` supports long strings and symbols